+++ resolwe/flow/exceptions.py
@@ -1,0 +1,7 @@
+"""Custom exceptions."""
+
+
+class SlugError(Exception):
+    """Slug error."""
+
+    pass