+++ docs/CHANGELOG.rst
@@ -27,6 +27,7 @@ Added
 Fixed
 -----
 - Prevent mutation of ``input_`` parameter in ``ProcessTestCase.run_process``
+- Return 400 instead of 500 error when slug already exists
 
 
 ===================